Fix x64 compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@20012 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseShell/PreserveChdir.cpp
@@ -24,7 +24,7 @@ PreserveChdir::PreserveChdir() :
     originalCurrentDirectory(size)
 {
     if (size > 0)
-        if (GetCurrentDirectory(size, originalCurrentDirectory) !=0)
+        if (GetCurrentDirectory((DWORD)size, originalCurrentDirectory) !=0)
             return; // succeeded
 
     // GetCurrentDirectory failed
